fix implicit conversion

+++ src/txn/txn_recover.c
@@ -487,19 +487,19 @@ __recovery_set_checkpoint_snapshot(WT_RECOVERY *r)
         WT_ERR_NOTFOUND_OK(
           __wt_config_getones(session, sys_config, WT_SYSTEM_CKPT_SNAPSHOT_MIN, &cval), false);
         if (cval.len != 0) {
-            conn->txn_global.snapshot_min = cval.val;
+            conn->txn_global.snapshot_min = (uint16_t)cval.val;
         }
 
         WT_ERR_NOTFOUND_OK(
           __wt_config_getones(session, sys_config, WT_SYSTEM_CKPT_SNAPSHOT_MAX, &cval), false);
         if (cval.len != 0) {
-            conn->txn_global.snapshot_max = cval.val;
+            conn->txn_global.snapshot_max = (uint16_t)cval.val;
         }
 
         WT_ERR_NOTFOUND_OK(
           __wt_config_getones(session, sys_config, WT_SYSTEM_CKPT_SNAPSHOT_COUNT, &cval), false);
         if (cval.len != 0) {
-            conn->txn_global.snapshot_count = cval.val;
+            conn->txn_global.snapshot_count = (uint16_t)cval.val;
         }
 
         WT_ERR_NOTFOUND_OK(
